Refactored some viewmodel classes into further subdivided namespaces.
