+++ FocalPointDMSClient/ViewModels/MainView/CustomerVm/CustomerMainViewModel.cs
@@ -1,10 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using System.Windows;
 
 using FocalPointDMSClient.Models.OrmModels;
 
-namespace FocalPointDMSClient.ViewModels.MainView
+namespace FocalPointDMSClient.ViewModels.MainView.CustomerVm
 {
     class CustomerMainViewModel : MainViewModel
     {
@@ -14,6 +15,9 @@ namespace FocalPointDMSClient.ViewModels.MainView
             RecordDeleteCommand = new DeleteCustomerCommand();
             RecordAddCommand = new NewCustomerCommand();
             EntityType = EntityType.Customer;
+
+            // Sets the MainViewModel application property to this particular instance.
+            Application.Current.Properties["MainViewModel"] = this;
         }
     }
 }
