+++ FocalPointDMSClient/ViewModels/MainView/CustomerVm/GetCustomerCommand.cs
@@ -1,16 +1,14 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Windows;
 using System.Windows.Input;
 
+using FocalPointDMSClient.Models.EntityConverters;
 using FocalPointDMSClient.ViewModels.CustomerDetailViewModel;
 using FocalPointDMSClient.Views;
-using FocalPointDMSClient.Models.OrmModels;
 
-namespace FocalPointDMSClient.ViewModels.MainView
+namespace FocalPointDMSClient.ViewModels.MainView.CustomerVm
 {
-    class NewCustomerCommand : ICommand
+    class GetCustomerCommand : ICommand
     {
         public bool CanExecute(object parameter)
         {
@@ -25,12 +23,14 @@ namespace FocalPointDMSClient.ViewModels.MainView
 
         public void Execute(object parameter)
         {
+            var mainViewModel = (MainViewModel)Application.Current.Properties["mainViewModel"];
+            var customerConverter = new CustomerConverter(mainViewModel.SelectedItemRow);
 
-            CustomerDetailVm viewModel = new CustomerDetailVm(new Customer());
+            var customer = customerConverter.Convert();
+
+            CustomerDetailVm viewModel = new CustomerDetailVm(customer);
             CustomerDetail customerDetail = new CustomerDetail(viewModel);
             customerDetail.Show();
-
-
         }
     }
 }
