+++ FocalPointDMSClient/ViewModels/MainView/GetCustomersCommand.cs
@@ -5,8 +5,9 @@ using System.Windows.Input;
 using FocalPointDMSClient.Models.DataTableBuilders;
 using FocalPointDMSClient.Models.OrmModels;
 using FocalPointDMSClient.ViewModels.MainView;
+using FocalPointDMSClient.ViewModels.MainView.CustomerVm;
 
-namespace FocalPointDMSClient.ViewModels
+namespace FocalPointDMSClient.ViewModels.MainView
 {
     public class GetCustomersCommand : ICommand
     {
@@ -23,7 +24,9 @@ namespace FocalPointDMSClient.ViewModels
 
         public void Execute(object parameter)
         {
-            var mainViewModel = (MainViewModel) Application.Current.Properties["mainViewModel"];
+            var mainViewModel = new CustomerMainViewModel();
+            Application.Current.Properties["mainViewModel"] =  mainViewModel;
+
             var factory = (DataTableBuilderFactory) Application.Current.Properties["DataTableBuilderFactory"];
             var controller = factory.GetInstance(EntityType.Customer);
             mainViewModel.MainDataTable = controller.BuildTable();
